autoformat {xtable}	R Documentation

Automatically Format Export Tables

Description

Suggest an appropriate alignment, number of digits, and display type for xtable.

Usage

autoformat(xtab, zap = getOption("digits"))

xalign(x, pad = TRUE)
xdigits(x, pad = TRUE, zap = getOption("digits"))
xdisplay(x, pad = TRUE)

Arguments

xtab
an object of class xtable.

x
a vector, matrix, or data frame.

pad
whether to format row names, when x is two-dimensional.

zap
the number of digits passed to zapsmall.

Value

autoformat returns a copy of xtab, after applying xalign, xdigits, and xdisplay.

xalign returns a character vector consisting of "l" and "r" elements, for left/right alignment.

xdigits returns an integer vector.

xdisplay returns a character vector of "d", "f", and "s" elements, for integer/double/string display.

Author(s)

Arni Magnusson.

See Also

xtable, align, digits, display

Examples

## 1  Vector
xalign(precip)
xdigits(precip)
xdisplay(precip)

## 2  Data frame
head(mtcars)
xdigits(mtcars, pad = FALSE)
xdigits(mtcars, pad = TRUE)
xalign(mtcars)
xdisplay(mtcars)

## 3  Autoformat when xtable is created
xtable(mtcars, align = xalign(mtcars), digits = xdigits(mtcars),
       display = xdisplay(mtcars))

## equivalent shortcut
xtable(mtcars, auto = TRUE)

## 4  Autoformat existing xtable
mt <- xtable(mtcars)
align(mt) <- xalign(mt)
digits(mt) <- xdigits(mt)
display(mt) <- xdisplay(mt)

## equivalent shortcut
mt <- autoformat(mt)

[Package xtable version 1.8-4 Index]