table.attributes {xtable}	R Documentation

Retrieve and Set Options for Export Tables

Description

Functions retrieving or setting table attributes for the supplied object of class "xtable".

Usage

  caption(x,...)
  caption(x) <- value
  label(x,...)
  label(x) <- value
  align(x,...)
  align(x) <- value
  digits(x,...)
  digits(x) <- value
  display(x,...)
  display(x) <- value

Arguments

x
An "xtable" object.

value
The value of the corresponding attribute.

...
Additional arguments. (Currently ignored.)

Details

These functions retrieve or set table attributes of the object x of class "xtable". See xtable for a description of the options.

Author(s)

David Dahl dahl@stat.tamu.edu with contributions and suggestions from many others (see source code).

See Also

xtable, print.xtable, formatC

[Package xtable version 1.7-1 Index]