tli {xtable}	R Documentation

Math scores from Texas Assessment of Academic Skills (TAAS)

Description

This data set contains math scores and demographic data of 100 randomly selected students participating in the Texas Assessment of Academic Skills (TAAS).

Usage

data(tli)

Format

A data.frame containing 100 observations with the following columns:

grade

Year in school of student

sex

Gender of student

disadvg

Is the student economically disadvantaged?

ethnicty

Race of student

tlimth

Math score of student

Source

Texas Education Agency, http://www.tea.state.tx.us

[Package xtable version 1.8-4 Index]