xtableMatharray {xtable}	R Documentation

Create LaTeX Mathematical Array

Description

Convert an array of numbers or mathematical expressions into an xtableMatharray object so it can be printed. A convenience function to enable the printing of arrays in mathematical expressions in LaTeX

Usage

xtableMatharray(x, caption = NULL, label = NULL, align = NULL,
                digits = NULL, display = NULL, auto = FALSE, ...)

Arguments

x
A numeric or character matrix.

caption
Character vector of length 1 or 2 containing the table's caption or title. If length is 2, the second item is the "short caption" used when LaTeX generates a "List of Tables". Set to NULL to suppress the caption. Default value is NULL. Included here only for consistency with xtable methods. Not expected to be of use.

label
Character vector of length 1 containing the LaTeX label. Set to NULL to suppress the label. Default value is NULL.

align
Character vector of length equal to the number of columns of the resulting table, indicating the alignment of the corresponding columns. Also, "|" may be used to produce vertical lines between columns in LaTeX tables, but these are effectively ignored when considering the required length of the supplied vector. If a character vector of length one is supplied, it is split as strsplit(align, "")[[1]] before processing. Since the row names are printed in the first column, the length of align is one greater than ncol(x) if x is a data.frame. Use "l", "r", and "c" to denote left, right, and center alignment, respectively. Use "p{3cm}" etc. for a LaTeX column of the specified width. For HTML output the "p" alignment is interpreted as "l", ignoring the width request. Default depends on the class of x.

digits
Numeric vector of length equal to one (in which case it will be replicated as necessary) or to the number of columns of the resulting table or matrix of the same size as the resulting table, indicating the number of digits to display in the corresponding columns. Since the row names are printed in the first column, the length of the vector digits or the number of columns of the matrix digits is one greater than ncol(x) if x is a data.frame. Default depends on the class of x. If values of digits are negative, the corresponding values of x are displayed in scientific format with abs(digits) digits.

display
Character vector of length equal to the number of columns of the resulting table, indicating the format for the corresponding columns. Since the row names are printed in the first column, the length of display is one greater than ncol(x) if x is a data.frame. These values are passed to the formatC function. Use "d" (for integers), "f", "e", "E", "g", "G", "fg" (for reals), or "s" (for strings). "f" gives numbers in the usual xxx.xxx format; "e" and "E" give n.ddde+nn or n.dddE+nn (scientific format); "g" and "G" put x[i] into scientific format only if it saves space to do so. "fg" uses fixed format as "f", but digits as number of significant digits. Note that this can lead to quite long result strings. Default depends on the class of x.

auto
Logical, indicating whether to apply automatic format when no value is passed to align, digits, or display. This ‘autoformat’ (based on xalign, xdigits, and xdisplay) can be useful to quickly format a typical matrix or data.frame. Default value is FALSE.

...
Additional arguments. (Currently ignored.)

Details

This function is only usable for production of LaTeX documents, not HTML.

Creates an object of class c("xtableMatharray","xtable","data.frame"), to ensure that it is printed by the print method print.xtableMatharray.

Value

An object of class c("xtableMatharray","xtable","data.frame").

Author(s)

David Scott <d.scott@auckland.ac.nz>

See Also

print.xtableMatharray

Examples

V <- matrix(c(1.140380e-03,  3.010497e-05,  7.334879e-05,
              3.010497e-05,  3.320683e-04, -5.284854e-05,
              7.334879e-05, -5.284854e-05,  3.520928e-04), nrow = 3)
mth <- xtableMatharray(V)
class(mth)
str(mth)
unclass(mth)

[Package xtable version 1.8-3 Index]