string {xtable}	R Documentation

String handling functions

Description

Private functions for conveniently working with strings.

Usage

  string(text,file="",append=FALSE)
  ## S3 method for class 'string'
print(x,...)
  ## S3 method for class 'string'
x + y
  as.string(x,file="",append=FALSE)
  is.string(x)

Arguments

text
A character object.

file
Name of the file that should receive the printed string.

append
Should the printed string be appended to the file?

x
A string object.

y
A string object.

...
Additional arguments. (Currently ignored.)

Details

These functions are private functions used by print.xtable. They are not intended to be used elsewhere.

Author(s)

David Dahl dahl@stat.byu.edu with contributions and suggestions from many others (see source code).

See Also

print.xtable

[Package xtable version 1.8-3 Index]